Update dimer geometry generation script to handle multiple conformers.
Conformers for a particular monomer should be given in a single .xyz file. The
script enforces a few restrictions on input files (compared to a normal .xyz
format), namely that conformers must have the same number of atoms, and must
have the elements in the same order, within each .xyz file. The update is
backwards compatible, and xyz files with a single conformer are handled
correctly.

Algorithmically, a confomer is selected at random from each Monomer A and
Monomer B with a uniform distribution each time a dimer is created. Otherwise
the dimer generation algorithms are unchanged.